move make_policy to config handler
